I added some changes to make the generated classes compile properly without having to be manually reordered.
Summary:

* When the 'generate' button is clicked, each class is added to a Class Dependency Graph, and each class that class references is added as a depenedency.
* There are two dependency types - POINTER and INSTANCE.
    * POINTER references can be solved with a forward declaration.  However, a simple reordering can solve that dependency and limit the forward declaration spam at the top of the generated file.
    * INSTANCE references can only be solved with reordering.  If class B contains an instance of class A, A must come before B in the generated file, no way around it.
* The graph is then recursively traversed - starting at nodes with no dependencies - with classes being added to the 'ordered' class collection.
* Circular dependencies (A points to B points to C points to A) are solved with forward declarations (C comes before B comes before A, with A being forward declared)
* The result is the output has fewer forward declarations and classes don't need to be manually reordered.  In my DayZ Reclass file (which inspired this work, because it takes me a lot of time manually reordering every time I generate the output) there are only 5 forward declarations with 356 classes.

Details:

The graph class itself is pretty simple.  It maintains a mapping of CNodeClass pointers to DependencyNodes.  Each node knows what class it represents, and has two vectors of edges - incoming edges (another class depends on this one) and outgoing edges (this class depends on another).  Each edge knows what node it points to, as well as whether the dependency is of POINTER or INSTANCE type.

Edges are generated by walking each node in a class and adding a dependency edge for each pointer, instance, instance array, and pointer array.  Parallel edges (duplicates) are not added, nor are simple recursive edges (a class has a pointer to an instance of that same class).  Other types (function ptrs?  Can we prototype those in ReClass?) may be necessary here, but these were all I could figure out without asking someone.

ClassDependencyGraph            depGraph;
    // Add each class as a node to the graph before adding dependency edges
    for (auto cNode : this->m_Classes) {
        depGraph.AddNode(cNode);
    }
    for (auto cNode : this->m_Classes) {
        for (size_t n = 0; n < cNode->NodeCount(); n++)
        {
            CNodeBase* pNode = (CNodeBase*)cNode->GetNode(n);
            NodeType Type = pNode->GetType();
            switch (Type) {
            case(nt_pointer):
            {
                CNodePtr* pPointer = (CNodePtr*)pNode;
                CNodeClass* pointerClass = pPointer->GetClass();
                depGraph.AddEdge(cNode, pointerClass, DependencyType::POINTER);
                break;
            }
            case(nt_instance):
            {
                CNodeClassInstance* pClassInstance = (CNodeClassInstance*)pNode;
                CNodeClass* instanceClass = pClassInstance->GetClass();
                depGraph.AddEdge(cNode, instanceClass, DependencyType::INSTANCE);
                break;
            }

            case(nt_array):
            {
                CNodeArray* pArray = (CNodeArray*)pNode;
                CNodeClass* instanceClass = pArray->GetClass();
                depGraph.AddEdge(cNode, instanceClass, DependencyType::INSTANCE);
                break;
            }

            case(nt_ptrarray):
            {
                CNodePtrArray* pArray = (CNodePtrArray*)pNode;
                CNodeClass* pointerClass = pArray->GetClass();
                depGraph.AddEdge(cNode, pointerClass, DependencyType::POINTER);
                break;
            }
            }
        }
    }

Ordering dependencies in the graph is simple in naive cases.  Just start at nodes that have no dependencies (leaf nodes) and recursively visit dependencies, adding them to an ordered class vector as you go.  This works great until you hit circular dependencies - you'd recurse until death.  That's where forward declarations are needed.  If you hit a node that is already being processed, you know you've thrown that ass in a circle.  So that's when you throw your hands up and forward declare - but you can only do that for a POINTER dependency, not an INSTANCE one.

You can see in this graph that DayZPlayer instances Entity which points to physicsObject which points to DayZPlayer.  And if that's the order you hit the nodes in, you're fine - just forward declare DayZPlayer.  However, there exists a path from gpWorld -> World -> N0000120E -> Entity -> PhysicsObject -> DayZPlayer -> Entity.  In that case, when you realize you're in a cycle you're at an INSTANCE dependency and you can't solve it with a forward declaration.

I wasn't sure what the correct way to handle this was, so I kind of improvised a solution.  At the beginning of graph solving, all nodes which have incoming INSTANCE type dependencies are marked.  Then, when they're visited during recursion, if they're being visited across a POINTER dependency edge, they're forward declared and not processed further.  So when you hit gpWorld -> World -> N0000120E -> Entity, it knows Entity has incoming INSTANCE dependencies, so instead of recursing and ending up in DayZPlayer, it adds a forward declaration for Entity and bails out.  Then, later (and along a different path that doesn't go through Entity), DayZPlayer is visited normally and it recurses into Entity, and all is well.

With forward declarations generated and the classes in dependency order, output goes on as normal, except instead of iterating of m_Classes, it iterates over the set of forward declarations and the in-order vector of classes instead.

To generate the visualized graph, I put a ToDot method in the DependencyGraph, which generates a GraphViz Dot representation of the graph.  I left it in as it may end up being useful in troubleshooting in the future.